Utils/curriculum.py
import os
import torch
import numpy as np
from collections import defaultdict
from torch.utils.data import Dataset, Subset, DataLoader
from PIL import Image
import torchvision.transforms as transforms
from tqdm import tqdm

class CurriculumLearningManager:
    """
    Implements curriculum learning for Halton-MaskGIT by gradually increasing
    the complexity of training samples based on various complexity metrics.
    """
    def __init__(self, dataset, complexity_metric='token_entropy', 
                 num_stages=5, batch_size=16, num_workers=4):
        """
        Args:
            dataset: The token dataset to apply curriculum learning to
            complexity_metric: Metric to determine complexity ('token_entropy', 'class_difficulty', 'edge_density')
            num_stages: Number of curriculum stages
            batch_size: Batch size for complexity analysis
            num_workers: Number of workers for data loading
        """
        self.dataset = dataset
        self.complexity_metric = complexity_metric
        self.num_stages = num_stages
        self.batch_size = batch_size
        self.num_workers = num_workers
        
        self.complexity_scores = None
        self.stage_indices = []
        self.current_stage = 0
        
    def analyze_dataset_complexity(self):
        """
        Analyze the entire dataset and assign complexity scores to each sample.
        """
        print(f"Analyzing dataset complexity using {self.complexity_metric}...")
        
        loader = DataLoader(
            self.dataset, 
            batch_size=self.batch_size,
            shuffle=False,
            num_workers=self.num_workers
        )
        
        complexity_scores = []
        
        for batch_idx, batch in enumerate(tqdm(loader)):
            # Tokens should be of shape [batch_size, height, width]
            tokens = batch[0] if isinstance(batch, tuple) else batch
            
            # Calculate complexity based on selected metric
            batch_scores = self.calculate_complexity(tokens)
            complexity_scores.extend(batch_scores)
        
        self.complexity_scores = np.array(complexity_scores)
        
        # Create curriculum stages based on complexity
        self._create_stages()
        
        print(f"Complexity analysis complete. Created {self.num_stages} curriculum stages.")
        return self.complexity_scores
    
    def calculate_complexity(self, tokens):
        """
        Calculate complexity scores for a batch of token samples.
        """
        if self.complexity_metric == 'token_entropy':
            # Calculate entropy of token distributions
            return self._calculate_token_entropy(tokens)
        
        elif self.complexity_metric == 'class_difficulty':
            # Use pre-defined class difficulty (e.g., from ImageNet confusion matrices)
            return self._calculate_class_difficulty(tokens)
        
        elif self.complexity_metric == 'edge_density':
            # Use edge density as a proxy for visual complexity
            return self._calculate_edge_density(tokens)
        
        else:
            raise ValueError(f"Unknown complexity metric: {self.complexity_metric}")
    
    def _calculate_token_entropy(self, tokens):
        """
        Calculate entropy of token distributions as a complexity measure.
        Higher entropy = more complex image.
        """
        batch_entropies = []
        
        for idx in range(tokens.shape[0]):
            # Get token counts
            token_sample = tokens[idx].flatten()
            unique, counts = torch.unique(token_sample, return_counts=True)
            probs = counts.float() / token_sample.numel()
            
            # Calculate entropy: -sum(p * log(p))
            entropy = -torch.sum(probs * torch.log2(probs + 1e-10)).item()
            batch_entropies.append(entropy)
            
        return batch_entropies
    
    def _calculate_class_difficulty(self, tokens):
        """
        Use class labels as a proxy for difficulty.
        This is a placeholder - in a real implementation, you would use 
        class difficulty metrics from literature or model performance.
        """
        # Extract class labels from filenames or metadata
        # For this example, we'll use random scores as placeholder
        return np.random.random(tokens.shape[0])
    
    def _calculate_edge_density(self, tokens):
        """
        Estimate visual complexity by measuring token transitions.
        More transitions = more complex image.
        """
        batch_edge_densities = []
        
        for idx in range(tokens.shape[0]):
            token_grid = tokens[idx]
            
            # Count horizontal transitions
            h_transitions = (token_grid[:, 1:] != token_grid[:, :-1]).float().sum()
            
            # Count vertical transitions
            v_transitions = (token_grid[1:, :] != token_grid[:-1, :]).float().sum()
            
            # Normalize by possible transitions
            total_possible = 2 * token_grid.numel() - token_grid.shape[0] - token_grid.shape[1]
            edge_density = (h_transitions + v_transitions).item() / total_possible
            
            batch_edge_densities.append(edge_density)
            
        return batch_edge_densities
    
    def _create_stages(self):
        """
        Divide dataset into stages based on complexity scores.
        """
        # Sort indices by complexity
        sorted_indices = np.argsort(self.complexity_scores)
        
        # Create stages with increasing complexity
        stage_size = len(sorted_indices) // self.num_stages
        
        self.stage_indices = []
        for i in range(self.num_stages):
            if i < self.num_stages - 1:
                stage_indices = sorted_indices[i * stage_size:(i + 1) * stage_size]
            else:
                # Last stage includes remaining samples
                stage_indices = sorted_indices[i * stage_size:]
                
            self.stage_indices.append(stage_indices)
    
    def get_current_stage_dataset(self):
        """
        Get a dataset subset for the current curriculum stage.
        """
        if self.complexity_scores is None:
            self.analyze_dataset_complexity()
            
        # For stage 0, return only the simplest examples
        # For later stages, include all examples up to current stage
        indices = []
        for i in range(self.current_stage + 1):
            indices.extend(self.stage_indices[i])
            
        return Subset(self.dataset, indices)
    
    def advance_stage(self):
        """
        Advance to the next curriculum stage if possible.
        Returns True if advanced, False if already at final stage.
        """
        if self.current_stage < self.num_stages - 1:
            self.current_stage += 1
            print(f"Advancing to curriculum stage {self.current_stage+1}/{self.num_stages}")
            return True
        return False
    
    def get_stage_info(self):
        """
        Get information about the current curriculum stage.
        """
        total_samples = sum(len(stage) for stage in self.stage_indices)
        current_samples = sum(len(self.stage_indices[i]) for i in range(self.current_stage + 1))
        
        return {
            "current_stage": self.current_stage + 1,
            "total_stages": self.num_stages,
            "samples_in_stage": len(self.stage_indices[self.current_stage]),
            "total_samples": total_samples,
            "current_samples": current_samples,
            "percentage": current_samples / total_samples * 100
        }

# Example integration with training loop
def curriculum_learning_example():
    """
    Example of how to integrate curriculum learning with training.
    """
    # Assume dataset is already created
    dataset = None  # Replace with actual dataset
    
    # Create curriculum learning manager
    curriculum = CurriculumLearningManager(
        dataset, 
        complexity_metric='token_entropy',
        num_stages=5
    )
    
    # Analyze dataset and create stages
    curriculum.analyze_dataset_complexity()
    
    # Training loop with curriculum
    num_epochs = 100
    curriculum_epochs = num_epochs // curriculum.num_stages
    
    for epoch in range(num_epochs):
        # Check if we should advance to the next curriculum stage
        if epoch > 0 and epoch % curriculum_epochs == 0:
            curriculum.advance_stage()
            
        # Get dataset for current stage
        stage_dataset = curriculum.get_current_stage_dataset()
        
        # Create data loader for current stage
        loader = DataLoader(
            stage_dataset,
            batch_size=16,
            shuffle=True,
            num_workers=4
        )
        
        # Print stage info
        stage_info = curriculum.get_stage_info()
        print(f"Epoch {epoch+1}/{num_epochs} - Stage {stage_info['current_stage']}/{stage_info['total_stages']} "
              f"({stage_info['percentage']:.1f}% of data)")
        
        # Training code would go here
        
    print("Training with curriculum learning complete!")
