requirements.txt
# Core Dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0

# 3D Processing
trimesh>=3.9.0
open3d>=0.15.1
pyvista>=0.32.0

# Attention and Transformers
einops>=0.6.1
xformers>=0.0.17

# Data Processing
tqdm>=4.64.0
Pillow>=9.0.0
PyYAML>=6.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
h5py>=3.6.0

# For ULIP-2
transformers>=4.20.0
sentence-transformers>=2.2.0

# Evaluation Metricstorch-fidelity>=0.3.0
lpips>=0.1.4

# Visualization
matplotlib>=3.5.0
plotly>=5.6.0
kaleido>=0.2.1

# Development
pytest>=7.0.0
black>=22.3.0
isort>=5.10.1
flake8>=4.0.0
